Index + Search Page

+++ src/main/java/it/univaq/we/internshipTutor/controller/InternshipController.java
@@ -2,10 +2,7 @@ package it.univaq.we.internshipTutor.controller;
 
 
 import it.univaq.we.internshipTutor.model.*;
-import it.univaq.we.internshipTutor.service.CompanyService;
-import it.univaq.we.internshipTutor.service.InternshipService;
-import it.univaq.we.internshipTutor.service.StudentInternshipService;
-import it.univaq.we.internshipTutor.service.UserService;
+import it.univaq.we.internshipTutor.service.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,8 +14,12 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import static it.univaq.we.internshipTutor.model.Popup.*;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import javax.validation.Valid;
+import java.security.Principal;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
 
 
@@ -36,6 +37,12 @@ public class InternshipController {
 
     @Autowired
     UserService userService;
+
+    @Autowired
+    ProfessorService professorService;
+
+    @Autowired
+    StudentService studentService;
 
     @RequestMapping(value={"/admin/create/internship"}, method = RequestMethod.POST)
     public String doCreate(@Valid @ModelAttribute("internship") Internship internship, BindingResult result, RedirectAttributes redirectAttributes) {
@@ -122,7 +129,8 @@ public class InternshipController {
     }
 
     @RequestMapping(value={"/company/create/internship"}, method = RequestMethod.POST)
-    public String doCreateByCompany(@Valid @ModelAttribute("internship") Internship internship, BindingResult result, RedirectAttributes redirectAttributes) {
+    public String doCreateByCompany(@Valid @ModelAttribute("internship") Internship internship, BindingResult result,
+                                    RedirectAttributes redirectAttributes, Principal principal) {
 
         if (result.hasErrors()) {
             // if there are errors during the binding (e.g. NotNull, Min, etc.)
@@ -140,7 +148,9 @@ public class InternshipController {
             return "redirect:/company/create/internship";
         }
 
-        //TODO assign to company_id the id of the company that is creating the intership
+        Company company = userService.findUserByEmail(principal.getName()).getCompany();
+
+        internship.setCompany(company);
 
         try{
             internshipService.save(internship);
@@ -161,7 +171,7 @@ public class InternshipController {
 
     @RequestMapping(value={"/company/delete/internship/{id}"}, method = RequestMethod.POST)
     public String doDeleteByCompany(@PathVariable(value = "id") Long id,
-                                    RedirectAttributes redirectAttributes) {
+                                    RedirectAttributes redirectAttributes, Principal principal) {
 
         if (id == null || id < 0) {
             // if there are errors during the binding (e.g. NotNull, Min, etc.)
@@ -171,8 +181,12 @@ public class InternshipController {
             return "redirect:/company/update/internship/" + id;
         }
 
-        //TODO check that the compnay is updating is own internship
-        //TODO force the values of "active" and "company" to be the same as they were before the update
+        Company company = userService.findUserByEmail(principal.getName()).getCompany();
+
+        // check that the company is deleting its own internship
+        if(!(company.getId().equals(internshipService.findInternshipById(id).getCompany().getId()))){
+            return "redirect:/error?code=403";
+        }
 
         try{
             internshipService.deleteInternshipById(id);
@@ -190,7 +204,8 @@ public class InternshipController {
     }
 
     @RequestMapping(value={"/company/update/internship"}, method = RequestMethod.POST)
-    public String doUpdateByCompany(@Valid @ModelAttribute("internship") Internship internship, BindingResult result, RedirectAttributes redirectAttributes) {
+    public String doUpdateByCompany(@Valid @ModelAttribute("internship") Internship internship, BindingResult result,
+                                    RedirectAttributes redirectAttributes, Principal principal) {
 
         if (result.hasErrors()) {
             // if there are errors during the binding (e.g. NotNull, Min, etc.)
@@ -201,8 +216,33 @@ public class InternshipController {
             return "redirect:/company/update/internship/" + internship.getId();
         }
 
-        //TODO check that the compnay is updating is own internship
-        //TODO force the values of "active" and "company" to be the same as they were before the update
+        Company company = userService.findUserByEmail(principal.getName()).getCompany();
+
+        Internship oldInternship = internshipService.findInternshipById(internship.getId());
+
+        // check that the company is updating its own internship
+        if(!(company.getId().equals(oldInternship.getCompany().getId()))){
+            return "redirect:/error?code=403";
+        }
+
+        if(!(company.getId().equals(internship.getCompany().getId()))){
+            return "redirect:/error?code=403";
+        }
+
+        // check that the values of "active" and "company" are the same as they were before the update
+        if(!(oldInternship.getActive().equals(internship.getActive()))){
+            redirectAttributes.addFlashAttribute("popup", new Popup("warning", "Cannot change Active attribute!"));
+            redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.internship", result);
+            redirectAttributes.addFlashAttribute("internship", internship);
+            return "redirect:/company/update/internship/" + internship.getId();
+        }
+
+        if(!(oldInternship.getCompany().getId().equals(internship.getCompany().getId()))){
+            redirectAttributes.addFlashAttribute("popup", new Popup("warning", WAR_MSG_EN));
+            redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.internship", result);
+            redirectAttributes.addFlashAttribute("internship", internship);
+            return "redirect:/company/update/internship/" + internship.getId();
+        }
 
         try{
             internshipService.save(internship);
@@ -219,13 +259,63 @@ public class InternshipController {
     }
 
 
-    @RequestMapping(value={"/company/create/internship"}, method = RequestMethod.GET)
-    public String renderCreateByCompany(ModelMap model, Pageable pageable) {
+    @RequestMapping(value = {"/internship/{id}"}, method = RequestMethod.POST)
+    public String doApplyInternship(@ModelAttribute("studentInternship") StudentInternship studentInternship, BindingResult result,
+                                    @PathVariable(value = "id") Long id, RedirectAttributes redirectAttributes, Principal principal) {
 
+        //if (result.hasErrors()) {
+        //    // if there are errors during the binding (e.g. NotNull, Min, etc.)
+        //    // redirect to the form displaying the errors
+        //    System.out.println("#####     Bindingresult errors:     #####");
+        //    List<FieldError> errors = result.getFieldErrors();
+        //    for (FieldError error : errors ) {
+        //        System.out.println (error.getField() + " - " + error.getDefaultMessage());
+        //    }
+        //    redirectAttributes.addFlashAttribute("popup", new Popup("warning", WAR_MSG_EN));
+        //    redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.internship", result);
+        //    redirectAttributes.addFlashAttribute("studentInternship", studentInternship);
+        //    return "redirect:/internship/" + id;
+        //}
+
+        Student currStudent = userService.findUserByEmail(principal.getName()).getStudent();
+
+        Internship currInternship = internshipService.findInternshipById(id);
+
+        List<StudentInternship> internships = studentInternshipService.findStudentInternshipByStudentAndInternship(currStudent, currInternship);
+
+        if(!internships.isEmpty()){
+            // Student already applied for this internship
+            redirectAttributes.addFlashAttribute("popup", new Popup("info", "You already applied for this Internship!"));
+            redirectAttributes.addFlashAttribute("org.springframework.validation.BindingResult.internship", result);
+            redirectAttributes.addFlashAttribute("studentInternship", studentInternship);
+            return "redirect:/internship/" + id;
+        }
+
+        // Insert the application
+        studentInternship.setStudent(currStudent);
+        studentInternship.setInternship(currInternship);
+        studentInternship.setAccepted(new Boolean(false));
+        studentInternship.setRejected(new Boolean(false));
+        studentInternship.setCompleted(new Boolean(false));
+
+        try{
+            studentInternshipService.save(studentInternship);
+            redirectAttributes.addFlashAttribute("popup", new Popup("success", "Application Sent Successfully!"));
+        }catch (Exception e){
+            e.printStackTrace();
+            redirectAttributes.addFlashAttribute("popup", new Popup("warning", WAR_MSG_EN_SAVE));
+            return "redirect:/internship/" + id;
+        }
+
+        return "redirect:/internship/" + id;
+    }
+
+
+    @RequestMapping(value={"/company/create/internship"}, method = RequestMethod.GET)
+    public String renderCreateByCompany(ModelMap model, Pageable pageable, Principal principal) {
 
         if(!model.containsAttribute("internship")){
-            //TODO set the company by looking at the one is session
-            Long userId = 47L;
+            Long userId = userService.findUserByEmail(principal.getName()).getId();
             Internship internship = new Internship(UUID.randomUUID());
             internship.setCompany(userService.findUserById(userId).getCompany());
             model.addAttribute("internship", internship);
@@ -242,7 +332,14 @@ public class InternshipController {
 
 
     @RequestMapping(value={"/company/update/internship/{id}"}, method = RequestMethod.GET)
-    public String renderUpdateByCompany(ModelMap model, Pageable pageable, @PathVariable(value = "id") Long id) {
+    public String renderUpdateByCompany(ModelMap model, Pageable pageable, @PathVariable(value = "id") Long id, Principal principal) {
+
+        Company company = userService.findUserByEmail(principal.getName()).getCompany();
+
+        // check that the company is updating its own internship
+        if(!(company.getId().equals(internshipService.findInternshipById(id).getCompany().getId()))){
+            return "redirect:/error?code=403";
+        }
 
         Internship i = internshipService.findInternshipById(id);
 
@@ -311,4 +408,46 @@ public class InternshipController {
         return "report";
     }
 
+    @RequestMapping(value = {"/internship/{id}"}, method = RequestMethod.GET)
+    public String renderInternship(ModelMap model, @PathVariable(value = "id") Long id) {
+
+        if(!model.containsAttribute("studentInternship")){
+            model.addAttribute("studentInternship", new StudentInternship(UUID.randomUUID()));
+        }
+        List<Professor> professors = professorService.findAll();
+        Internship internship = internshipService.findInternshipById(id);
+        User userCompany = userService.findUserByCompany(internship.getCompany().getId());
+        Company company = internship.getCompany();
+
+        model.addAttribute("professors", professors);
+        model.addAttribute("internship", internship);
+        model.addAttribute("user", userCompany);
+        model.addAttribute("company", company);
+
+        return "internship";
+    }
+
+    @RequestMapping(value = {"/internships"}, method = RequestMethod.GET)
+    public String renderInternship(ModelMap model, @RequestParam(value = "s", required = false) String query, Pageable pageable){
+
+        Map<Long, User> pippo = new HashMap<>();
+        Page<Internship> internships;
+
+        if(query != null){
+            internships = internshipService.findIntershipsByQuery(pageable, query);
+        } else {
+            internships = internshipService.findAll(pageable);
+        }
+
+        for(Internship internship : internships){
+            pippo.put(internship.getId(), userService.findUserByCompany(internship.getCompany().getId()));
+        }
+
+        model.addAttribute("pippo", pippo);
+
+        PageWrapper<Internship> page = new PageWrapper<>(internships, "/internship");
+        model.addAttribute("internships", internships.getContent());
+        model.addAttribute("page", page);
+        return "search";
+    }
 }
